% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartostamp-package.R
\docType{package}
\name{quartostamp-package}
\alias{quartostamp}
\alias{quartostamp-package}
\title{quartostamp: RStudio Addin to Insert Quarto Formatting Blocks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

RStudio Addin to help write Quarto documents by inserting common formatting blocks such as columns, tabsets, callouts, incremental lists and Revealjs speaker notes. User-selected text is automatically wrapped, otherwise skeleton examples are inserted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matt-dray.github.io/quartostamp/}
  \item \url{https://github.com/matt-dray/quartostamp}
  \item Report bugs at \url{https://github.com/matt-dray/quartostamp/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

Other contributors:
\itemize{
  \item Zoë Turner (\href{https://orcid.org/0000-0003-1033-9158}{ORCID}) [contributor]
}

}
\keyword{internal}
