% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_bluecoloredspan}
\alias{stamp_bluecoloredspan}
\title{Insert Blue Colored Span}
\usage{
stamp_bluecoloredspan()
}
\description{
Insert a span with css color specification with default color "red
}
\details{
Output looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[]\{style='color:blue;'\}
}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
