% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_divider}
\alias{stamp_divider}
\title{Insert Slide Divider}
\usage{
stamp_divider()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a slide divider in a Revealjs presentation (for slides without
titles).
}
\details{
The output looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{---

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-markers:
\code{\link[=stamp_pause]{stamp_pause()}}
}
\concept{stamp-markers}
