% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_pause}
\alias{stamp_pause}
\title{Insert Pause}
\usage{
stamp_pause()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a pause marker into a Revealjs presentation slide.
}
\details{
The output looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{. . .

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-markers: 
\code{\link{stamp_divider}()}
}
\concept{stamp-markers}
