% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_redcoloredspan}
\alias{stamp_redcoloredspan}
\title{Insert Red Colored Span}
\usage{
stamp_redcoloredspan()
}
\description{
Insert a span with css color specification with default color "red
}
\details{
Output looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[]\{style='color:red;'\}
}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
