% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartostamp-package.R
\docType{package}
\name{quartostamp-package}
\alias{quartostamp}
\alias{quartostamp-package}
\title{quartostamp: RStudio Addin to Insert Quarto Helpers}
\description{
An RStudio Addin to insert ('stamp') pre-written classes and divs into your Quarto documents.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
