% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartostamp-package.R
\docType{package}
\name{quartostamp-package}
\alias{quartostamp}
\alias{quartostamp-package}
\title{quartostamp: Insert Quarto Helpers}
\description{
An RStudio Addin to help you insert ('stamp') useful little code blocks into your Quarto document.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
