% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartostamp-package.R
\docType{package}
\name{quartostamp-package}
\alias{quartostamp}
\alias{quartostamp-package}
\title{quartostamp: RStudio Addin to Insert Quarto Helpers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An RStudio Addin to insert ('stamp') pre-written classes and divs into your Quarto documents. URL: https://matt-dray.github.io/quartostamp/, https://github.com/matt-dray/quartostamp
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/matt-dray/quartostamp/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
