% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_aside}
\alias{stamp_aside}
\title{Insert Aside}
\usage{
stamp_aside()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a fenced div that creates an 'aside' in a Revealjs presentation slide.
Will embed text selected by the user, otherwise skeleton help text will be
inserted.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: aside
Additional commentary.
:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-blocks: 
\code{\link{stamp_columns}()},
\code{\link{stamp_footer}()},
\code{\link{stamp_notes}()},
\code{\link{stamp_tabset}()}
}
\concept{stamp-blocks}
