% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_aside}
\alias{stamp_aside}
\title{Insert Aside}
\usage{
stamp_aside()
}
\description{
Insert a div that creates an 'aside' in a Revealjs presentation slide.
}
\details{
Output looks like this:\preformatted{::: aside
Additional commentary.
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
