% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_aside}
\alias{stamp_aside}
\title{Insert Aside}
\usage{
stamp_aside()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a div that creates an 'aside' in a Revealjs presentation slide. Will
embed text selected by the user, otherwise skeleton help text will be
inserted.
}
\details{
The output looks like this if the user hadn't selected any text:\preformatted{::: aside
Additional commentary.
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
