% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_caution}
\alias{stamp_callout_caution}
\title{Insert 'Caution' Callout Block}
\usage{
stamp_callout_caution()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a callout-block fenced div of type 'caution', which has an orange
accent and a traffic-cone icon. Will embed text selected by the user,
otherwise skeleton help text will be inserted.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{:::\{.callout-caution collapse=false appearance='default' icon=true\}
## Optional caption (caution)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-callouts:
\code{\link[=stamp_callout_important]{stamp_callout_important()}},
\code{\link[=stamp_callout_note]{stamp_callout_note()}},
\code{\link[=stamp_callout_tip]{stamp_callout_tip()}},
\code{\link[=stamp_callout_warning]{stamp_callout_warning()}}
}
\concept{stamp-callouts}
