% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_caution}
\alias{stamp_callout_caution}
\title{Insert 'Caution' Callout Block}
\usage{
stamp_callout_caution()
}
\description{
Insert a callout-block div of type 'caution', which has an orange accent and
a traffic-cone icon.
}
\details{
Output looks like this:\preformatted{:::\{.callout-caution collapse=false appearance='default' icon=true\}
## Optional caption (caution)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
