% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_important}
\alias{stamp_callout_important}
\title{Insert 'Important' Callout Block}
\usage{
stamp_callout_important()
}
\description{
Insert a callout-block div of type 'important', which has a red accent and
an exclamation-point icon.
}
\details{
Output looks like this:\preformatted{:::\{.callout-important collapse=false appearance='default' icon=true\}
## Optional caption (important)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
