% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_note}
\alias{stamp_callout_note}
\title{Insert 'Note' Callout Block}
\usage{
stamp_callout_note()
}
\description{
Insert a callout-block div of type 'note', which has a blue accent and
a letter 'i' icon.
}
\details{
Output looks like this:\preformatted{:::\{.callout-note collapse=false appearance='default' icon=true\}
## Optional caption (note)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
