% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_note}
\alias{stamp_callout_note}
\title{Insert 'Note' Callout Block}
\usage{
stamp_callout_note()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a callout-block div of type 'note', which has a blue accent and
a letter 'i' icon. Will embed text selected bythe  user, otherwise
skeleton help text will be inserted.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{:::\{.callout-note collapse=false appearance='default' icon=true\}
## Optional caption (note)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
