% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_tip}
\alias{stamp_callout_tip}
\title{Insert 'Tip' Callout Block}
\usage{
stamp_callout_tip()
}
\description{
Insert a callout-block div of type 'tip', which has an green accent and
a lightbulb icon.
}
\details{
Output looks like this:\preformatted{:::\{.callout-tip collapse=false appearance='default' icon=true\}
## Optional caption (tip)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
