% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_callout_warning}
\alias{stamp_callout_warning}
\title{Insert 'Warning' Callout Block}
\usage{
stamp_callout_warning()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a callout-block div of type 'warning', which has a yellow accent and
a warning-triangle icon. Will embed text selected by the user, otherwise
skeleton help text will be inserted.
}
\details{
The output looks like this if the user hadn't selected any text:\preformatted{:::\{.callout-warning collapse=false appearance='default' icon=true\}
## Optional caption (warning)
- Hide callout body with 'collapse=true'
- Remove 'collapse' to prevent expandability
- Set appearance to 'default', 'simple' or 'minimal'
- Remove icon with 'icon=false'
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
