% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_columns}
\alias{stamp_columns}
\title{Insert Column Layout}
\usage{
stamp_columns()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert afenced  div for a horizontal two-column layout with percentage widths
in a Revealjs presentation slide. Will embed text selected by the user into
the left column and helper text into the right, otherwise skeleton help text
will be inserted into both columns.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{:::: \{.columns\}

::: \{.column width='40\%'\}
Left column
:::

::: \{.column width='60\%'\}
Right column
:::

::::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-blocks: 
\code{\link{stamp_aside}()},
\code{\link{stamp_footer}()},
\code{\link{stamp_notes}()},
\code{\link{stamp_tabset}()}
}
\concept{stamp-blocks}
