% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_columns}
\alias{stamp_columns}
\title{Insert Column Layout}
\usage{
stamp_columns()
}
\description{
Insert a div for a horizontal two-column layout with percentage widths in a
Revealjs presentation slide.
}
\details{
Output looks like this:\preformatted{:::: \{.columns\}

::: \{.column width='40\%'\}
Left column
:::

::: \{.column width='60\%'\}
Right column
:::

::::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
