% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_footer}
\alias{stamp_footer}
\title{Insert Custom Footer}
\usage{
stamp_footer()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a custom footer div in a Revealjs presentation slide. Will include
text selected by the user, otherwise skeleton help text will be inserted.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: footer
Custom footer text
:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-blocks:
\code{\link[=stamp_aside]{stamp_aside()}},
\code{\link[=stamp_columns]{stamp_columns()}},
\code{\link[=stamp_notes]{stamp_notes()}},
\code{\link[=stamp_tabset]{stamp_tabset()}}
}
\concept{stamp-blocks}
