% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_footer}
\alias{stamp_footer}
\title{Insert Custom Footer}
\usage{
stamp_footer()
}
\description{
Insert a custom footer div in a Revealjs presentation slide.
}
\details{
Output looks like this:\preformatted{::: footer
Custom footer text
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
