% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_footnote}
\alias{stamp_footnote}
\title{Insert Footnote}
\usage{
stamp_footnote()
}
\description{
Insert a footnote marker.
}
\details{
Output looks like this:\preformatted{^[A footnote]
}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
