% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_incremental}
\alias{stamp_incremental}
\title{Insert Incremental List}
\usage{
stamp_incremental()
}
\description{
Insert a list-containing div in a Revealjs presentation slide that is
revealed incrementally.
}
\details{
Output looks like this:\preformatted{::: \{.incremental\}
- List element A
- List element B
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
