% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_incremental}
\alias{stamp_incremental}
\title{Insert Incremental List}
\usage{
stamp_incremental()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a list-containing div in a Revealjs presentation slide that is
revealed incrementally. Will embed text selected by the user, otherwise
skeleton help text will be inserted.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: \{.incremental\}
- List element A
- List element B
:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
