% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_nonincremental}
\alias{stamp_nonincremental}
\title{Insert Non-Incremental List}
\usage{
stamp_nonincremental()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a list-containing fenced div to a Revealjs presentation slide that is
not revealed incrementally  (overrides a global incremental reveal). Will
embed text selected by the user, otherwise skeleton help text will be
inserted.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: \{.nonincremental\}
- List element A
- List element B
:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-lists:
\code{\link[=stamp_incremental]{stamp_incremental()}}
}
\concept{stamp-lists}
