% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_nonincremental}
\alias{stamp_nonincremental}
\title{Insert Non-incremental List}
\usage{
stamp_nonincremental()
}
\description{
Insert a list to a Revealjs presentation slide that is not revealed
incrementally  (to override global incremental reveal).
}
