% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_nonincremental}
\alias{stamp_nonincremental}
\title{Insert Non-Incremental List}
\usage{
stamp_nonincremental()
}
\description{
Insert a list-containing div to a Revealjs presentation slide that is not
revealed incrementally  (overrides a global incremental reveal).
}
