% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_notes}
\alias{stamp_notes}
\title{Insert Speaker Notes}
\usage{
stamp_notes()
}
\description{
Insert a speaker notes div in a Revealjs presentation slide.
}
\details{
Output looks like this:\preformatted{::: \{.notes\}
Speaker notes (press 's' when presenting to switch to speaker mode).
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
