% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_notes}
\alias{stamp_notes}
\title{Insert Speaker Notes}
\usage{
stamp_notes()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a speaker notes div in a Revealjs presentation slide. Will embed
text selected by the user, otherwise skeleton help text will be inserted.
}
\details{
The output looks like this if the user hadn't selected any text:\preformatted{::: \{.notes\}
Speaker notes (press 's' when presenting to switch to speaker mode).
:::

}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
