% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_scrollable}
\alias{stamp_scrollable}
\title{Insert Scrollable Content}
\usage{
stamp_scrollable()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a class that allows content to be scrolled if the slide is too small
to hold it (place on the same line as the slide title).
}
\details{
Output looks like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\{.scrollable\}
}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
\seealso{
Other stamp-classes:
\code{\link[=stamp_smaller]{stamp_smaller()}}
}
\concept{stamp-classes}
