% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_scrollable}
\alias{stamp_scrollable}
\title{Insert Scrollable Content}
\usage{
stamp_scrollable()
}
\description{
Insert a class that allows content to be scrolled if the slide is too small
to hold it.
}
