% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_smaller}
\alias{stamp_smaller}
\title{Insert Smaller Font}
\usage{
stamp_smaller()
}
\description{
Insert a class that makes the font smaller in a Revealjs presentation slide.
(place on the same line as the slide title).
}
