% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_smaller}
\alias{stamp_smaller}
\title{Insert Smaller Font}
\usage{
stamp_smaller()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a class that makes the font smaller in a Revealjs presentation slide
(place on the same line as the slide title).
}
\details{
Output looks like this:\preformatted{\{.smaller\}
}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
