% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_tabset}
\alias{stamp_tabset}
\title{Insert Tabset}
\usage{
stamp_tabset()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a panel tabset div to a Revealjs presentation slide. Will embed text
selected by the user into the first tab and skeleton help text into the
second, otherwise skeleton help text will be inserted into both tabs.
}
\details{
The output looks like this if the user hadn't selected any text:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: \{.panel-tabset\}

## Tab A

Content for Tab A

## Tab B

Content for Tab B

:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
