% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamps.R
\name{stamp_tabset}
\alias{stamp_tabset}
\title{Insert Tabset}
\usage{
stamp_tabset()
}
\value{
Nothing. Text is updated in the active document.
}
\description{
Insert a panel tabset div. Will embed text selected by the user into the
first tab and skeleton help text into the second, otherwise skeleton help
text will be inserted into both tabs. The header level will depend on the
current document format. Tabs default to header level 2, but will be header
level 3 if the format in the YAML header is specified as 'revealjs'. This is
because a level 2 header demarcates slide boundaries in a reveal.js
presentation written with Quarto.
}
\details{
The output looks like this if the user hadn't selected any text and the
document format is not 'revealjs' according to the YAML header:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: \{.panel-tabset\}

## Tab A

Content for Tab A

## Tab B

Content for Tab B

:::

}\if{html}{\out{</div>}}

And if the document has format 'revealjs' in the YAML header:

\if{html}{\out{<div class="sourceCode">}}\preformatted{::: \{.panel-tabset\}

### Tab A

Content for Tab A

### Tab B

Content for Tab B

:::

}\if{html}{\out{</div>}}
}
\references{
\href{https://quarto.org/docs/reference/}{The Quarto documentation website.}
}
